fix: deep audit — unused imports, event listener leak, code dedup, better dedupe
Bugs found and fixed from deep audit of all 18 PRs merged today:

1. StatsCollector: removed unused imports (Account, AccountKind, config)
   and dead event listener that leaked (registered but never removed)

2. Webhook trade dedup: improved inner instruction duplicate detection
   to compare slab_address + size (not just trader + side) to prevent
   missing trades with same direction in one tx

3. Code dedup: extracted decodeBase58, readU128LE, parseTradeSize to
   shared utils/binary.ts — was duplicated across webhook.ts and
   TradeIndexer.ts (130 lines removed)

4. Removed unused eventBus import from StatsCollector
